ai_anim_pipeline/requirements.txt
pyyaml
pydantic
loguru
rich
pillow
